Highlight goal progress on lunch-in notification
