+++ app/src/main/res/layout/fragment_my_goal.xml
@@ -45,26 +45,17 @@
 
         <TextView
             android:id="@+id/fragment_my_goal_progress_gained"
-            android:layout_width="match_parent"
+            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="20dp"
             android:textSize="16sp"
             />
         <TextView
             android:id="@+id/fragment_my_goal_progress_remaining"
-            android:layout_width="match_parent"
+            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="20dp"
             android:text="Progress remaining: $2202"
-            android:textSize="16dp"
-            />
-
-        <TextView
-            android:id="@+id/fragment_my_goal_other"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="20dp"
-            android:text="Team - Feel free to add content or give me ideas!"
             android:textSize="16dp"
             />
